move to a flat structure
