--- packages/fields/src/actions/base-post/index.ts
@@ -1,6 +1,0 @@
-export { default as viewPost } from './view-post';
-export { default as reorderPage } from './reorder-page';
-export { default as reorderPageNative } from './reorder-page.native';
-export { default as duplicatePost } from './duplicate-post';
-export { default as duplicatePostNative } from './duplicate-post.native';
-export { default as renamePost } from './rename-post';
