--- packages/fields/src/actions/common/index.ts
@@ -1,5 +1,0 @@
-export { default as viewPostRevisions } from './view-post-revisions';
-export { default as permanentlyDeletePost } from './permanently-delete-post';
-export { default as restorePost } from './restore-post';
-export { default as trashPost } from './trash-post';
-export { default as deletePost } from './delete-post';
