+++ packages/fields/src/actions/duplicate-pattern.tsx
@@ -9,8 +9,8 @@ import type { Action } from '@wordpress/dataviews';
 /**
  * Internal dependencies
  */
-import { unlock } from '../../lock-unlock';
-import type { Pattern } from '../../types';
+import { unlock } from '../lock-unlock';
+import type { Pattern } from '../types';
 
 // Patterns.
 const { CreatePatternModalContents, useDuplicatePatternProps } =
