--- packages/fields/src/actions/pattern/index.ts
@@ -1,3 +1,0 @@
-export { default as duplicatePattern } from './duplicate-pattern';
-export { default as exportPattern } from './export-pattern';
-export { default as exportPatternNative } from './export-pattern.native';
