--- packages/fields/src/actions/template/index.ts
@@ -1,1 +1,0 @@
-export { default as resetTemplateAction } from './reset-template';
